requirements_clean.txt
streamlit==1.28.1
numpy==1.24.3
scikit-learn==1.3.1
joblib==1.3.2
Pillow==10.0.0
